chore:  optimize Skeleton sub component
